May 13th 2020 Changes
Corrected data simulation code: 

Changed the categorical variable simulation lines 34-40 to simulate k dummy variables for a categorical covariate with k+1 levels. 

Swapped the columns in "mats" to match the number of cases and controls with and without minor alleles, as described in the actual simulation analysis example. 

Thanks to Y. Guo for catching these errors.